# userauth: don't register no-op auth functions and unify public API