--- a/sqlite3_test.go
+++ b/sqlite3_test.go
@@ -2112,6 +2112,7 @@ var benchmarks = []testing.InternalBenchmark{
 	{Name: "BenchmarkRows", F: benchmarkRows},
 	{Name: "BenchmarkStmtRows", F: benchmarkStmtRows},
 	{Name: "BenchmarkQueryParallel", F: benchmarkQueryParallel},
+	{Name: "BenchmarkOpen", F: benchmarkOpen},
 }
 
 func (db *TestDB) mustExec(sql string, args ...any) sql.Result {
@@ -2585,4 +2586,19 @@ func benchmarkQueryParallel(b *testing.B) {
 			}
 		}
 	})
+}
+
+func benchmarkOpen(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		db, err := sql.Open("sqlite3", ":memory:")
+		if err != nil {
+			b.Fatal(err)
+		}
+		if err := db.Ping(); err != nil {
+			b.Fatal(err)
+		}
+		if err := db.Close(); err != nil {
+			b.Fatal(err)
+		}
+	}
 }