This commit changes Open to not register any of the auth functions when
userauth is disabled. It also unifies the public API of this library for
when userauth is enabled/disabled.

**NOTE:** This PR builds on and includes https://github.com/mattn/go-sqlite3/pull/1302

```
goos: darwin
goarch: arm64
pkg: github.com/charlievieth/go-sqlite3
cpu: Apple M4 Pro
                       │   y1.txt    │               y3.txt                │
                       │   sec/op    │   sec/op     vs base                │
Suite/BenchmarkOpen-14   13.76µ ± 1%   12.37µ ± 1%  -10.11% (p=0.000 n=10)

                       │   y1.txt   │               y3.txt               │
                       │    B/op    │    B/op     vs base                │
Suite/BenchmarkOpen-14   712.0 ± 0%   111.0 ± 1%  -84.41% (p=0.000 n=10)

                       │   y1.txt    │               y3.txt               │
                       │  allocs/op  │ allocs/op   vs base                │
Suite/BenchmarkOpen-14   20.000 ± 0%   3.000 ± 0%  -85.00% (p=0.000 n=10)
```