@charlievieth 2025-01-13
@mattn / @rittneje could either of you take a look at https://github.com/mattn/go-sqlite3/pull/1302 and this (this PR depends on https://github.com/mattn/go-sqlite3/pull/1302)?