@rittneje 2025-04-05
@mattn I think #1302 is supposed to go first, as these changes include those changes.