@charlievieth 2025-04-08
> @mattn I think https://github.com/mattn/go-sqlite3/pull/1302 is supposed to go first, as these changes include those changes.

Correct, we should ideally merge https://github.com/mattn/go-sqlite3/pull/1302 before this PR (it's not strictly necessary since this PR contains the commit https://github.com/mattn/go-sqlite3/pull/1308/commits/83213c4e1a9696d43e32eac967f776b50d54ed41 from [PR](https://github.com/mattn/go-sqlite3/pull/1302) but for the sake of documentation and having commits tied to GH PRs it would probably best to merge that one first).

Thank you both for taking the time to review this!